refactor(core): 将 delta_time 类型从 float 改为 double
- 在 GameApp 和 Time 类中，将 delta_time 相关的函数和变量类型从 float 改为 double
- 这个改动可以提高时间计算的精度，避免在长时间运行或高帧率情况下出现精度问题
- 需要相应地修改使用这些函数的其他部分，以确保类型一致性和避免类型转换带来的潜在问题

+++ src/engine/core/game_app.cpp
@@ -1,7 +1,7 @@
 /***
  * @Author: hanmi255 hanmi2550505@gmail.com
  * @Date: 2025-07-29 14:34:14
- * @LastEditTime: 2025-07-31 11:46:09
+ * @LastEditTime: 2025-07-31 12:39:15
  * @LastEditors: hanmi255 hanmi2550505@gmail.com
  * @Description:
  * @FilePath: \SunnyLand\src\engine\core\game_app.cpp
@@ -37,7 +37,7 @@ namespace engine::core {
 
         while (is_running_) {
             time_->update();
-            float delta_time = time_->getDeltaTime();
+            double delta_time = time_->getDeltaTime();
 
             handleEvents();
             update(delta_time);
@@ -77,7 +77,7 @@ namespace engine::core {
         }
     }
 
-    void GameApp::update(float /* delta_time */)
+    void GameApp::update(double /* delta_time */)
     {
         testCamera();
     }
